+++ __tests__/i18n-5.ts
@@ -1,4 +1,5 @@
 import assert from 'assert';
+
 import { i18n, changeCurrentLang } from '../test-fixtures/temp/i18n-5';
 
 test('i18n-5: function params', async () => {
