+++ __tests__/i18n-6.ts
@@ -1,4 +1,5 @@
 import assert from 'assert';
+
 import { i18n, changeCurrentLang } from '../test-fixtures/temp/i18n-6';
 
 test('i18n-6: tricky arguments', async () => {
