+++ package.json
@@ -16,7 +16,7 @@
   },
   "prettier": "@willbooster/prettier-config",
   "dependencies": {
-    "lodash": "4.17.20",
+    "lodash-es": "4.17.15",
     "yargs": "16.2.0"
   },
   "devDependencies": {
@@ -28,7 +28,7 @@
     "@babel/preset-env": "7.12.10",
     "@babel/preset-typescript": "7.12.7",
     "@types/jest": "26.0.19",
-    "@types/lodash": "4.14.165",
+    "@types/lodash-es": "4.17.4",
     "@types/node": "14.14.13",
     "@types/yargs": "15.0.11",
     "@typescript-eslint/eslint-plugin": "4.9.1",
